py/requirements.txt
xmltodict
requests
mysql.connector-python